+++ connect/amazon/dynamodb/fake/src/test/kotlin/org/http4k/connect/amazon/dynamodb/LocalDynamoTest.kt
@@ -3,7 +3,7 @@ package org.http4k.connect.amazon.dynamodb
 import org.http4k.client.JavaHttpClient
 import org.http4k.connect.amazon.dynamodb.model.TableName
 import org.http4k.connect.amazon.fakeAwsEnvironment
-import org.http4k.connect.assumeDockerDaemonRunning
+import org.http4k.util.assumeDockerDaemonRunning
 import org.http4k.core.Uri
 import org.http4k.core.then
 import org.http4k.filter.ClientFilters.SetBaseUriFrom