+++ connect/amazon/dynamodb/fake/src/testFixtures/kotlin/org/http4k/connect/amazon/dynamodb/DynamoDbSource.kt
@@ -1,7 +1,7 @@
 package org.http4k.connect.amazon.dynamodb
 
 import org.http4k.connect.amazon.fakeAwsEnvironment
-import org.http4k.connect.assumeDockerDaemonRunning
+import org.http4k.util.assumeDockerDaemonRunning
 import org.http4k.core.Uri
 import org.testcontainers.containers.GenericContainer