+++ connect/kafka/rest/client/src/test/kotlin/org/http4k/connect/kafka/rest/LocalKafkaRestTest.kt
@@ -1,7 +1,7 @@
 package org.http4k.connect.kafka.rest
 
 import org.http4k.client.JavaHttpClient
-import org.http4k.connect.assumeDockerDaemonRunning
+import org.http4k.util.assumeDockerDaemonRunning
 import org.http4k.connect.kafka.rest.v2.KafkaRestV2Contract
 import org.http4k.core.Uri
 import org.http4k.util.PortBasedTest