+++ connect/storage/redis/src/test/kotlin/org/http4k/connect/storage/RedisStorageTest.kt
@@ -8,7 +8,7 @@ import io.lettuce.core.codec.StringCodec
 import io.lettuce.core.protocol.CommandArgs
 import io.mockk.every
 import io.mockk.mockk
-import org.http4k.connect.assumeDockerDaemonRunning
+import org.http4k.util.assumeDockerDaemonRunning
 import org.http4k.core.Uri
 import org.junit.jupiter.api.Test
 import org.testcontainers.containers.GenericContainer